migrating back to how inmemory channels are in master.
